+++ src/error.rs
@@ -25,6 +25,13 @@ pub enum Error {
         /// Can be `None` when it is expected outside the current file
         expected_line: Option<NonZeroUsize>,
     },
+    /// A diagnostic code matcher was declared but had no matching error.
+    CodeNotFound {
+        /// The code that was not found, and the span of where that code was declared.
+        code: Spanned<String>,
+        /// Can be `None` when it is expected outside the current file
+        expected_line: Option<NonZeroUsize>,
+    },
     /// A ui test checking for failure does not have any failure patterns
     NoPatternsFound,
     /// A ui test checking for success has failure patterns