+++ src/tests.rs
@@ -32,6 +32,7 @@ fn main() {
                 message:"Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             }
         ]
     ];
@@ -68,6 +69,7 @@ fn main() {
                     message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                     level: Level::Error,
                     line_col: None,
+                    code: None,
                 }
             ]
         ];
@@ -94,6 +96,7 @@ fn main() {
                     message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                     level: Level::Error,
                     line_col: None,
+                    code: None,
                 }
             ]
         ];
@@ -124,6 +127,7 @@ fn main() {
                     message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                     level: Level::Note,
                     line_col: None,
+                    code: None,
                 }
             ]
         ];
@@ -164,6 +168,7 @@ fn main() {
                 message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             }
         ]
     ];
@@ -201,11 +206,13 @@ fn main() {
                 message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             },
             Message {
                 message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             }
         ]
     ];
@@ -249,16 +256,19 @@ fn main() {
                 message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             },
             Message {
                 message: "kaboom".to_string(),
                 level: Level::Warn,
                 line_col: None,
+                code: None,
             },
             Message {
                 message: "cake".to_string(),
                 level: Level::Warn,
                 line_col: None,
+                code: None,
             },
         ],
     ];
@@ -281,6 +291,7 @@ fn main() {
                     message,
                     level: Level::Warn,
                     line_col: _,
+                    code: None,
                 }] if message == "kaboom" => {}
                 _ => panic!("{:#?}", msgs),
             }
@@ -312,16 +323,19 @@ fn main() {
                 message: "Undefined Behavior: type validation failed: encountered a dangling reference (address 0x10 is unallocated)".to_string(),
                 level: Level::Error,
                 line_col: None,
+                code: None,
             },
             Message {
                 message: "kaboom".to_string(),
                 level: Level::Warn,
                 line_col: None,
+                code: None,
             },
             Message {
                 message: "cake".to_string(),
                 level: Level::Warn,
                 line_col: None,
+                code: None,
             },
         ],
     ];
@@ -338,5 +352,172 @@ fn main() {
     match &errors[..] {
         [] => {}
         _ => panic!("{:#?}", errors),
+    }
+}
+
+#[test]
+fn find_code() {
+    let s = r"
+fn main() {
+    let _x: i32 = 0u32; //~ E0308
+}
+    ";
+    let config = config();
+    let comments = Comments::parse(s, config.comment_defaults.clone(), Path::new("")).unwrap();
+    {
+        let messages = vec![
+            vec![],
+            vec![],
+            vec![],
+            vec![Message {
+                message: "mismatched types".to_string(),
+                level: Level::Error,
+                line_col: None,
+                code: Some("E0308".into()),
+            }],
+        ];
+        let mut errors = vec![];
+        check_annotations(
+            messages,
+            vec![],
+            Path::new("moobar"),
+            &mut errors,
+            "",
+            &comments,
+        )
+        .unwrap();
+        match &errors[..] {
+            [] => {}
+            _ => panic!("{:#?}", errors),
+        }
+    }
+
+    // different error code
+    {
+        let messages = vec![
+            vec![],
+            vec![],
+            vec![],
+            vec![Message {
+                message: "mismatched types".to_string(),
+                level: Level::Error,
+                line_col: None,
+                code: Some("SomeError".into()),
+            }],
+        ];
+        let mut errors = vec![];
+        check_annotations(
+            messages,
+            vec![],
+            Path::new("moobar"),
+            &mut errors,
+            "",
+            &comments,
+        )
+        .unwrap();
+        match &errors[..] {
+            [Error::CodeNotFound { code, .. }, Error::ErrorsWithoutPattern { msgs, .. }]
+                if **code == "E0308" && code.line().get() == 3 && msgs.len() == 1 => {}
+            _ => panic!("{:#?}", errors),
+        }
+    }
+
+    // warning instead of error
+    {
+        let messages = vec![
+            vec![],
+            vec![],
+            vec![],
+            vec![Message {
+                message: "mismatched types".to_string(),
+                level: Level::Warn,
+                line_col: None,
+                code: Some("E0308".into()),
+            }],
+        ];
+        let mut errors = vec![];
+        check_annotations(
+            messages,
+            vec![],
+            Path::new("moobar"),
+            &mut errors,
+            "",
+            &comments,
+        )
+        .unwrap();
+        match &errors[..] {
+            [Error::CodeNotFound { code, .. }] if **code == "E0308" && code.line().get() == 3 => {}
+            _ => panic!("{:#?}", errors),
+        }
+    }
+}
+
+#[test]
+fn find_code_with_prefix() {
+    let s = r"
+fn main() {
+    let _x: i32 = 0u32; //~ E0308
+}
+    ";
+    let mut config = config();
+    config.comment_defaults.base().diagnostic_code_prefix =
+        Spanned::dummy("prefix::".to_string()).into();
+    let comments = Comments::parse(s, config.comment_defaults.clone(), Path::new("")).unwrap();
+    {
+        let messages = vec![
+            vec![],
+            vec![],
+            vec![],
+            vec![Message {
+                message: "mismatched types".to_string(),
+                level: Level::Error,
+                line_col: None,
+                code: Some("prefix::E0308".into()),
+            }],
+        ];
+        let mut errors = vec![];
+        check_annotations(
+            messages,
+            vec![],
+            Path::new("moobar"),
+            &mut errors,
+            "",
+            &comments,
+        )
+        .unwrap();
+        match &errors[..] {
+            [] => {}
+            _ => panic!("{:#?}", errors),
+        }
+    }
+
+    // missing prefix
+    {
+        let messages = vec![
+            vec![],
+            vec![],
+            vec![],
+            vec![Message {
+                message: "mismatched types".to_string(),
+                level: Level::Error,
+                line_col: None,
+                code: Some("E0308".into()),
+            }],
+        ];
+        let mut errors = vec![];
+        check_annotations(
+            messages,
+            vec![],
+            Path::new("moobar"),
+            &mut errors,
+            "",
+            &comments,
+        )
+        .unwrap();
+        match &errors[..] {
+            [Error::CodeNotFound { code, .. }, Error::ErrorsWithoutPattern { msgs, .. }]
+                if **code == "prefix::E0308" && code.line().get() == 3 && msgs.len() == 1 => {}
+            _ => panic!("{:#?}", errors),
+        }
     }
 }